Class member affiliation and structure
Updated the ReadMe.
Made occational comments.
Created a new UmlClassDialog for creating and editing classes.
Created a UmlMember abstract class for the UmlMemberVar and
UmlMemberFunc classes to derive from.
Restructured the UmlMemberVar and UmlMemberFunc for the data moved into
UmlMember.
Made the UmlClass creation a two staged process, as seen in
UmlBlocks.cpp.
Created 'Create' function in UmlClass as second part of the two-staged
creation process.
Added content to the UmlClass size evaluation process, in the
UpdateShapeSize function.
Added content for the add variable/function functions for the UmlClass.
Changed the internal function and variable containers from std::vector
to std::list.

-Lex